% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{common_colors}
\alias{common_colors}
\title{Common hues from the color wheel}
\format{
## `common_colors`
A data frame with 12 rows and 2 columns:
\describe{
  \item{color}{Common color name}
  \item{hue}{Hue code between 0 and 360}
  ...
}
}
\source{
<https://coolors.co/>
}
\usage{
common_colors
}
\description{
A collection of the prominent hues from the color wheel at 30 degree intervals
and their common names at 30 degree intervals
}
\keyword{datasets}
