% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{contrast_text_color}
\alias{contrast_text_color}
\title{Make a color object}
\usage{
contrast_text_color(hex_code)
}
\arguments{
\item{hex_code}{The target color background}
}
\value{
A hexadecimal color code best for contrast with the specified color
background.
}
\description{
`contrast_text_color()` finds the best black/white text contrast for a
specified color background.
}
\examples{
# White background returns black
contrast_text_color("#FFFFFF")

# Dark red background returns white
contrast_text_color("#AC192D")
}
