% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{hsl_to_hex}
\alias{hsl_to_hex}
\title{HSL color to HEX color}
\usage{
hsl_to_hex(h, s, l)
}
\arguments{
\item{h}{An integer between 0 and 360, representing the color's "hue"}

\item{s}{An integer between 0 and 100, representing the color's "saturation"}

\item{l}{An integer between 0 and 100, representing the color's "lightness"}
}
\value{
A string representing the hexadecimal color code, prepended with a "#"
}
\description{
`hsl_to_hex()` turns a set of "hue", "saturation", and "lightness" values to
its equivalent hexadecimal code value.
}
\examples{
hsl_to_hex(115, 80, 50)
}
