% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{main_color_obj}
\alias{main_color_obj}
\title{Main color object}
\usage{
main_color_obj()
}
\value{
A list, or color object, featuring four named elements
("core', "background", "text", "lines")
}
\description{
`main_color_obj()` supplies a color object responsible for styling the
`theme_adamb()` ggplot theme layer when no other object is passed thereto.
}
\examples{
main_color_obj()
}
\seealso{
[theme_adamb()]
}
