% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{make_color_obj}
\alias{make_color_obj}
\title{Make a color object}
\usage{
make_color_obj(
  type = c("Light", "Dark"),
  hue = NULL,
  name = NULL,
  print = FALSE
)
}
\arguments{
\item{type}{The style of palette, either "Light" or "Dark"}

\item{hue}{An integer between 0 and 360, representing the target palette
color's "hue"}

\item{name}{A valid string representing a color name from `common_colors`}

\item{print}{Whether to print the palette to the console, TRUE or FALSE}
}
\value{
A list of named elements making up a color palette
}
\description{
`make_color_obj()` creates a color palette wrapped in a list of named elements from
a target hue or color name.
}
\examples{
# In it's simplest form, specify a type and a default hue or valid color name from `common_colors`:
make_color_obj("Light", hue = 10)
make_color_obj("Dark", name = "Orange")

# Additionally, you can opt to print the HSL values inspiring the palette to the console:
make_color_obj("Light", name = "Blue", print = TRUE)
}
