% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{mysql_close}
\alias{mysql_close}
\title{Close MySQL connection}
\usage{
mysql_close(cname = NULL)
}
\arguments{
\item{cname}{A string representing the connection variable name from the environment}
}
\value{
Nothing
}
\description{
`mysql_close()` closes an open MySQL server connection
}
\examples{
mysql_close("mycname")
}
\seealso{
[mysql_connect()]
}
