% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{mysql_peek_fields}
\alias{mysql_peek_fields}
\title{Peek MySQL database tables}
\usage{
mysql_peek_fields(cname = NULL, table = NULL)
}
\arguments{
\item{cname}{A string representing the connection variable name from the environment}

\item{table}{A string representing the table name for which to list fields}
}
\value{
A vector of table field names from the database connection
}
\description{
`mysql_peek_fields()` lists the field names present in a particular table
from a valid MySQL database connection
}
\examples{
mysql_peek_fields("mycname", "mytablename")

# If connection was made with default cname from `mysql_connect()`, it may be called
# with only the "table" parameter
mysql_peek_fields(table = "mytablename")
}
\seealso{
[mysql_connect(), mysql_peek_tables()]
}
