% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{mysql_peek_tables}
\alias{mysql_peek_tables}
\title{Peek MySQL database tables}
\usage{
mysql_peek_tables(cname = NULL)
}
\arguments{
\item{cname}{A string representing the connection variable name from the environment;}
}
\value{
A vector of database table names
}
\description{
`mysql_peek_tables()` lists the table names present from a valid MySQL database connection
}
\examples{
mysql_peek_tables("mycname")

# If connection was made with default cname from `mysql_connect()`, it may be called
# with no parameters
mysql_peek_tables()
}
\seealso{
[mysql_connect()]
}
