% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sql.R
\name{mysql_send_query}
\alias{mysql_send_query}
\title{Send MySQL query}
\usage{
mysql_send_query(cname = NULL, query = NULL)
}
\arguments{
\item{cname}{A string representing the connection variable name from the environment}

\item{query}{A string representing the query to execute}
}
\value{
A dataframe of the query results
}
\description{
`mysql_send_query()` sends a query to the database per the connection passed,
closes the connection when appropriate, and returns the result
}
\examples{
mysql_send_query("mycname", "SELECT * FROM mytable")
}
\seealso{
[mysql_connect(), mysql_peek_tables(), mysql_peek_fields()]
}
