% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{office_data}
\alias{office_data}
\title{The Office Sales data set}
\format{
## `office_data`
A data frame with 18 rows and 3 columns:
\describe{
  \item{month}{Full month name}
  \item{sales_person}{First name of sales person}
  \item{sales}{Rounded sales figure}
}
}
\source{
<https://www.peacocktv.com/stream-tv/the-office>
}
\usage{
office_data
}
\description{
Data set inspired by The Office hit TV show featuring sales figures displayed in
S07:E14 of the show.
}
\keyword{datasets}
