% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{scale_to_gradient}
\alias{scale_to_gradient}
\title{Scale a gradient}
\usage{
scale_to_gradient(x, len_x, min_x, max_x, colors)
}
\arguments{
\item{x}{The continuous variable to map}

\item{len_x}{The length of the data frame or vector corresponding to x}

\item{min_x}{The minimum value found in the variable "x"}

\item{max_x}{The maximum value found in the variable "x"}

\item{colors}{The target colors for the gradient}
}
\value{
A hexadecimal color code mapped from "x" to the graident
}
\description{
`scale_to_gradient()` will map a continuous variable to a custom gradient
}
\examples{
data <- data.frame(
  sales = rnorm(10, 50, 10)
)

data$color <- scale_to_gradient(
  sales, nrow(data), min(sales), max(sales), c("green", "red")
)
}
