% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{theme_adamb}
\alias{theme_adamb}
\title{Adam B ggplot2 theme}
\usage{
theme_adamb(color_obj = NULL)
}
\arguments{
\item{color_obj}{A list, featuring at least four elements, each
containing a hexadecimal color value: "core", "background", "text",
and "lines". May be ignored in which case the main color object will
be used.}
}
\value{
A theme element for laying onto a ggplot object.
}
\description{
`theme_adamb()` applies styling to ggplot objects per the preferences of
Adam B
}
\examples{
theme_adamb()

# Optionally, pass a color object, featuring at least four named elements
# ("core', "background", "text", "lines"), each featuring a hexadecimal color
# value:
theme_adamb(list(core = "#1b7e64", background = "#e4f1ed", text = "#24423a", lines = "#c7d1ce"))
}
\seealso{
[main_color_obj()]
}
