% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataviz.R
\name{make_pal}
\alias{make_pal}
\title{Make a color palette}
\usage{
make_pal(type = c("Light", "Dark"), hue = 110, name = NULL, print = FALSE)
}
\arguments{
\item{type}{The style of palette, either "Light" or "Dark"}

\item{hue}{An integer between 0 and 360, representing the target palette
color's "hue"}

\item{name}{A valid string representing a color name from `common_colors`}

\item{print}{Whether to print the palette to the console, TRUE or FALSE}
}
\value{
A list of named elements making up a color palette
}
\description{
`make_pal()` creates a color palette wrapped in a list of named elements from
a target hue or color name.
}
